requirements.txt
requests
trimesh
numpy
pillow
python-dotenv
open3d
matplotlib
tqdm
